# frontend/openapi.json
{"openapi": "3.1.0", "info": {"title": "AIVE Backend API", "description": "Modern AI Video Editor Backend - Interactive video creation process", "version": "0.1.0"}, "paths": {"/{full_path}": {"options": {"summary": "Options Handler", "description": "Handle CORS preflight requests.", "operationId": "options_handler__full_path__options", "parameters": [{"name": "full_path", "in": "path", "required": true, "schema": {"type": "string", "title": "Full Path"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/health": {"get": {"summary": "Health Check", "description": "Health check endpoint.", "operationId": "health_check_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/projects/": {"get": {"tags": ["Projects"], "summary": "Get All Projects", "description": "Get a list of all projects.", "operationId": "get_all_projects_api_v1_projects__get", "parameters": [{"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "Skip"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "default": 100, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"type": "object", "additionalProperties": true}, "title": "Response Get All Projects Api V1 Projects  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["Projects"], "summary": "Create Project", "description": "Create a new project with audio file upload, transcription, and footage recommendations.", "operationId": "create_project_api_v1_projects__post", "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_create_project_api_v1_projects__post"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Response Create Project Api V1 Projects  Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/{project_id}": {"get": {"tags": ["Projects"], "summary": "Get Project Details", "description": "Get details for a specific project.", "operationId": "get_project_details_api_v1_projects__project_id__get", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Response Get Project Details Api V1 Projects  Project Id  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["Projects"], "summary": "Update Project", "description": "Update a project by ID.", "operationId": "update_project_api_v1_projects__project_id__put", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Project Data"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Response Update Project Api V1 Projects  Project Id  Put"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "patch": {"tags": ["Projects"], "summary": "Patch Project", "description": "Partially update a project by ID.", "operationId": "patch_project_api_v1_projects__project_id__patch", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Project Data"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Response Patch Project Api V1 Projects  Project Id  Patch"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["Projects"], "summary": "Delete Project", "description": "Delete a project by ID.", "operationId": "delete_project_api_v1_projects__project_id__delete", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/{project_id}/footage": {"post": {"tags": ["Projects"], "summary": "Submit Footage Choices", "description": "Submit footage choices for sentences and get music recommendations.", "operationId": "submit_footage_choices_api_v1_projects__project_id__footage_post", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FootageChoices"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MusicResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/{project_id}/render": {"post": {"tags": ["Projects"], "summary": "Render Project", "description": "Start rendering a video for a project.", "operationId": "render_project_api_v1_projects__project_id__render_post", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "responses": {"202": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/render/status/{task_id}": {"get": {"tags": ["Projects"], "summary": "Get Project Render Status", "description": "Get the status of a render task (project-scoped route).", "operationId": "get_project_render_status_api_v1_projects_render_status__task_id__get", "parameters": [{"name": "task_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Task Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/projects/{project_id}/generate-title": {"post": {"tags": ["Projects"], "summary": "Generate Title For Project", "description": "Generate an AI-powered title for an existing project based on its content.", "operationId": "generate_title_for_project_api_v1_projects__project_id__generate_title_post", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Response Generate Title For Project Api V1 Projects  Project Id  Generate Title Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/render/{project_id}/render": {"post": {"tags": ["Render"], "summary": "Render Project", "description": "Start rendering a video for a project.", "operationId": "render_project_api_v1_render__project_id__render_post", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RenderRequest"}}}}, "responses": {"202": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RenderResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/render/status/{task_id}": {"get": {"tags": ["Render"], "summary": "Get Render Status", "description": "Get the status of a render task.", "operationId": "get_render_status_api_v1_render_status__task_id__get", "parameters": [{"name": "task_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Task Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RenderStatusResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["Render"], "summary": "Update Render Status", "description": "Update render task status (internal endpoint).", "operationId": "update_render_status_api_v1_render_status__task_id__put", "parameters": [{"name": "task_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Task Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Status Update"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": {"type": "string"}, "title": "Response Update Render Status Api V1 Render Status  Task Id  Put"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/render/{project_id}/tasks": {"get": {"tags": ["Render"], "summary": "Get Project Render Tasks", "description": "Get all render tasks for a project.", "operationId": "get_project_render_tasks_api_v1_render__project_id__tasks_get", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Response Get Project Render Tasks Api V1 Render  Project Id  Tasks Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"Body_create_project_api_v1_projects__post": {"properties": {"audio_file": {"type": "string", "format": "binary", "title": "Audio File"}}, "type": "object", "required": ["audio_file"], "title": "Body_create_project_api_v1_projects__post"}, "FootageChoice": {"properties": {"sentence_id": {"type": "string", "title": "Sentence Id"}, "footage_url": {"type": "string", "title": "Footage Url"}}, "type": "object", "required": ["sentence_id", "footage_url"], "title": "FootageChoice", "description": "Legacy footage choice schema."}, "FootageChoices": {"properties": {"footage_choices": {"items": {"$ref": "#/components/schemas/FootageChoice"}, "type": "array", "title": "Footage Choices"}}, "type": "object", "required": ["footage_choices"], "title": "FootageChoices", "description": "Legacy footage choices collection."}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "MusicRecommendation": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "url": {"type": "string", "title": "Url"}}, "type": "object", "required": ["id", "name", "url"], "title": "MusicRecommendation", "description": "Legacy music recommendation schema."}, "MusicResponse": {"properties": {"project_id": {"type": "string", "title": "Project Id"}, "recommended_music": {"items": {"$ref": "#/components/schemas/MusicRecommendation"}, "type": "array", "title": "Recommended Music"}}, "type": "object", "required": ["project_id", "recommended_music"], "title": "MusicResponse", "description": "Music recommendation response."}, "RenderRequest": {"properties": {"add_subtitles": {"type": "boolean", "title": "Add Subtitles", "default": true}, "include_audio": {"type": "boolean", "title": "Include Audio", "default": true}}, "type": "object", "title": "RenderRequest", "description": "Schema for video render requests."}, "RenderResponse": {"properties": {"render_task_id": {"type": "string", "title": "Render Task Id"}, "status_url": {"type": "string", "title": "Status Url"}}, "type": "object", "required": ["render_task_id", "status_url"], "title": "RenderResponse", "description": "Response schema for render requests."}, "RenderStatusResponse": {"properties": {"status": {"type": "string", "title": "Status"}, "video_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Video Url"}, "error": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Error"}, "progress": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Progress"}}, "type": "object", "required": ["status"], "title": "RenderStatusResponse", "description": "Response schema for render status queries."}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}